requirements.txt
pygame-ce==2.3.2  # Image opening and scanning